create new package, move DataException.java, DataReader.java, DataReaderTest.java
