+++ src/main/java/org/training/triangles/Director.java
@@ -1,5 +1,7 @@
 package org.training.triangles;
 
+import org.training.triangles.data.DataReader;
+import org.training.triangles.logic.DataException;
 import org.training.triangles.logic.TriangleLineValidator;
 import org.training.triangles.model.Triangle;
