+++ src/main/java/org/training/triangles/TriangleCreator.java
@@ -1,13 +1,12 @@
 package org.training.triangles;
 
+import org.training.triangles.logic.DataException;
 import org.training.triangles.model.Point;
 import org.training.triangles.model.Triangle;
 
-import java.util.List;
-
 public class TriangleCreator {
 
-    public Triangle createTriangle (String lineCoordinates) throws DataException{
+    public Triangle createTriangle (String lineCoordinates) throws DataException {
         String[] triangleCoordinates = lineCoordinates.split(" ");
         Point pointOne = new Point(Double.parseDouble(triangleCoordinates[0]), Double.parseDouble(triangleCoordinates[1]));
         Point pointTwo = new Point(Double.parseDouble(triangleCoordinates[0]), Double.parseDouble(triangleCoordinates[1]));
