+++ src/main/java/org/training/triangles/logic/DataException.java
@@ -1,4 +1,4 @@
-package org.training.triangles;
+package org.training.triangles.logic;
 
 public class DataException extends Exception {
     public DataException(String message, Throwable e) {
