+++ src/main/java/org/training/triangles/model/Triangle.java
@@ -1,11 +1,11 @@
 package org.training.triangles.model;
 
 public class Triangle {
-    private Point firstPoint;
+    private final Point firstPoint;
 
-    private Point secondPoint;
+    private final Point secondPoint;
 
-    private Point thirdPoint;
+    private final Point thirdPoint;
 
     public Triangle (Point firstPoint, Point secondPoint, Point thirdPoint) {
         this.firstPoint = firstPoint;
@@ -13,25 +13,16 @@ public class Triangle {
         this.thirdPoint = thirdPoint;
     }
 
-    public void setFirstPoint(Point setOnePoint) {
-        firstPoint = setOnePoint;
-    }
 
     public Point getFirstPoint() {
         return firstPoint;
     }
 
-    public void setSecondPoint(Point setSecondPoint) {
-        secondPoint = setSecondPoint;
-    }
 
     public Point getSecondPoint() {
         return secondPoint;
     }
 
-    public void setThirdPoint(Point setThirdPoint) {
-        thirdPoint = setThirdPoint;
-    }
 
     public Point getThirdPoint() {
