+++ src/test/java/org/training/triangles/DirectorTest.java
@@ -3,6 +3,8 @@ package org.training.triangles;
 import org.junit.Assert;
 import org.junit.Test;
 import org.mockito.Mockito;
+import org.training.triangles.data.DataReader;
+import org.training.triangles.logic.DataException;
 import org.training.triangles.logic.TriangleLineValidator;
 import org.training.triangles.model.Point;
 import org.training.triangles.model.Triangle;
